+++ examples/Ex_Poisson-AuxSourceTerm.py
@@ -1,7 +1,7 @@
 # %%
 from petsc4py import PETSc
 import underworld3 as uw
-from underworld3.poisson import Poisson
+from underworld3.systems import Poisson
 import numpy as np
 
 options = PETSc.Options()
@@ -17,7 +17,7 @@ options["snes_monitor_short"] = None
 options["snes_rtol"] = 1.0e-7
 
 # %%
-mesh = uw.Mesh(elementRes=(9,9), minCoords=(-2.2,-.4))
+mesh = uw.mesh.Mesh(elementRes=(9,9), minCoords=(-2.2,-.4))
 bnds = mesh.boundary
 
 # %%
@@ -35,7 +35,7 @@ y1 = mesh.maxCoords[1]
 y0 = mesh.minCoords[1]
 
 # %%
-diff = uw.MeshVariable( mesh=mesh, num_components=1, name="diff", vtype=uw.mesh.VarType.SCALAR, degree=u_degree )
+diff = uw.mesh.MeshVariable( mesh=mesh, num_components=1, name="diff", vtype=uw.mesh.VarType.SCALAR, degree=u_degree )
 
 # %%
 # example of setting the auxiliary field by numpy array, a.k.a by hand