+++ underworld3/__init__.py
@@ -1,2 +1,4 @@
-from .mesh import Mesh, MeshVariable, Spherical
-from .swarm import Swarm, SwarmVariable
+import underworld3.mesh
+import underworld3.maths
+import underworld3.swarm
+import underworld3.systems