+++ underworld3/systems/__init__.py
@@ -1,0 +1,2 @@
+from .stokes import Stokes
+from .poisson import Poisson